Detect same PIC Image and Reboot
If the PIC bootloader is entered via setting the DFM "UpdateAvailable" flag, and it is detected that the PIC OTA image is the same as the version currently running, then the PIC will boot the current image.

This is an edge case and should not be seen.  Test code uses the BLE System Control characteristic to force the condition (src/ble/src/bleInterface.c).

Replace the Image Download complete semaphore with a queued message.

TODO - Test normal update.
TODO - Remove semaphore code