Saving Host Ota Image State
Host Ota Image State was not being saved in NVM upon completion of an OTA Job( accepted or rejected).  Consequently the saved state remained in Pending and caused subsequent update jobs to be rejected.

Add a call to setImageState() from hostOta_setImageState() resolves the issue.

Additionally code was added to clear a latent error.  i.e. if the Host Ota Image State was already set to Pending in NVM, this had to be cleared upon receipt of a new OTA Job.  A new (non-Test) job, with the Image Status set to Pending, was already trapped and caused hostOta_ResetDevice() to be called.  This function is only called under error conditions.  Adding a call to hostOta_setImageState( eOTA_PAL_ImageState_Invalid ) clears the latent error.

In the above function call to esp_restart() was removed.  There is no reason to reset the ESP on a PIC update failure.